requirements.txt
Flask
boto>=2.19.0
requests
yapsy
pyOpenSSL